increase SharedInformer resync period to 10 minutes
Whenever SharedInformerFactories are specified a resync period
has to be specified. A resync events plays back all the events held
in the infromer cache. Since, The controller uses a workqueue
for processing and requeues any events that caused errors, This
resync period is largely not useful and is necessary to compensate
any bugs in the controller that are missed by the requeuing logic
in the workqueue.

This means that higher period is always advised, as its rare
to miss events. Hence, this period is being set to 10 minutes.

10 minutes is also the same period that the destination and other
controllers use in the core linkerd control-plane.

More on this: https://groups.google.com/g/kubernetes-sig-api-machinery/c/PbSCXdLDno0/m/v9gH3HXVDAAJ

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>