New SMI extension, which includes the `smi-adaptor` (#1)
This PR adds the new SMI Extension, which includes the
`CLI` and the `smi-adaptor`. Along with these components, 
scripts that are needed to build, run and test them are also
added. Patterns from the linkerd2 repo are followed, without
inheriting unnecessary complexity.

This PR also adds unit tests and integration tests for the same

## `smi-adaptor` controller

Runs in a loop and listens for `TrafficSplit` Resources
and makes sure that the relevant `ServiceProfile` state is present.
This follows the same pattern as that of the [`k8s/sample-controller`](https://github.com/kubernetes/sample-controller)

The state changes are as follows:

- User creates or updates a `TS` resource
    - No relevant `SP` already present
       - **Create New `SP`**
    - Relevant `SP` already present
      - `SP` contains the `smi.linkerd.io/skip` annotation
        - **Skip**
      - No `smi.linkerd.io/skip` Annotation
        - **Update `SP.dstOverrides` field**
- User deletes a `TS` resource
   - **empty `dstOverrides` field**

## Building and Running

First, Install a Linkerd enabled Kubernetes cluster.
Then, Run the following commands:
```bash
# Build the CLi
./bin/build-cli-bin
# Build the smi-adaptor docker image
./bin/docker-build
# Load the image into kind
kind load docker-image cr.l5d.io/linkerd/smi-adaptor:$(./bin/root-tag)
# Run the smi install cmd
./target/cli/linux-amd64/linkerd-smi install | k apply -f -
```

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>