Add `linkerd smi check` sub command
This PR adds a new check sub command which is used to perform
a check, and make sure that the SMI components are installed
and functioning correctly.

This follows the same pattern of check that we use in viz, jaeger, etc
other extensions.

This PR also updates the integration test to use the command, instead
of doing a `kubectl rollout status`

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>