readme: fix actions workflow badge
This PR updates the README to use the `integration_tests` workflow
for the status badge generation.

I tried to find if there was a way to show the status based on
multiple workflows to show accurate status (i.e `unit_tests`, etc in
our case) but GitHub doesn't seem to support that yet.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>