controller: add configuration to set number of workers (#5)
Currently, We only have a single goroutine i.e worker that processes
items from the workqueue. But there is nothing preventing us from
having multiple goroutines processing items from workqueues. and one
of the benefit of doing this whole workqueue way is to be able to
process concurrently.

This PR adds a new `worker-threads` CLI flag to controller, to override
the default number of goroutines i.e 2.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>