controller: use a struct as workqueue item
Currently, We use a string as the item for the workqueue, which
is encoded from a TS and decoded to a TS. This is done, instead
of passing the full TS into the workqueue and using it, because of
updates that could happen and hence the latest has to be fetched always
in `SyncHandler`.

This PR, removes the additional encoding and decoding work by using
a struct in its place, and thus removing the need for decoding.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>