Support multi-arch builds
This PR updates the `docker-build` and `build-cli-bin` scripts
to support multi-arch builds. These are enabled by using
`DOCKER_MULTIARCH` and `CLI_MULTIARCH` environment variables
respectively.

Currently, For the `build-cli-bin` only the `linux-amd64` is
built by default irrespective of the local system environment.
(Should we change it to be the local system's arch instead?)

**These env variables are not used yet, and will be used
in the release workflow, which will be a separate PR***

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>