Use helm chart version in helm integration tests (#53)
PR https://github.com/linkerd/linkerd-smi/pull/20 updated SMI so that the app version was distinct from the chart version.  This means that, as of this release, the chart version is 1.0.1 whereas the app version is v0.2.1.  However, the Helm integration test is still wired to use the app version and attempts to load the v0.2.1 of the Helm chart, which does not exist.

Update the Helm integration test to use the Helm chart version instead.  Note that this is currently hard coded in the release workflow and will need to be updated each time the Helm chart version is changed.

Signed-off-by: Alex Leong <alex@buoyant.io>